!feat: convert to ESM modules with strict TypeScript
- Convert to ES modules with .js extensions in imports
- Enable strict TypeScript with ES2022 target
- Add exactOptionalPropertyTypes and noUncheckedIndexedAccess
- Update all imports to use .js extensions
- Enable incremental builds for faster compilation

BREAKING CHANGE: Full ESM migration requires .js extensions in imports
