+++ __tests__/git.test.ts
@@ -1,5 +1,6 @@
-import * as git from '../src/git'
-import {ChangeStatus} from '../src/file'
+import { describe, test, expect } from 'vitest'
+import * as git from '../src/lib/git.js'
+import { ChangeStatus } from '../src/file.js'
 
 describe('parsing output of the git diff command', () => {
   test('parseGitDiffOutput returns files with correct change status', async () => {
